+++ tests/tests_main.c
@@ -9,20 +9,20 @@ int main(int argc, char** argv)
     palLog(nullptr, "%s: %s", "PAL Version", palGetVersionString());
 
     // core
-    registerTest("Logger Test", loggerTest);
-    registerTest("Time Test", timeTest);
-    registerTest("User Event Test", userEventTest);
-    registerTest("Event Test", eventTest);
+    // registerTest("Logger Test", loggerTest);
+    // registerTest("Time Test", timeTest);
+    // registerTest("User Event Test", userEventTest);
+    // registerTest("Event Test", eventTest);
 
 #if PAL_HAS_SYSTEM
-    registerTest("System Test", systemTest);
+    // registerTest("System Test", systemTest);
 #endif // PAL_HAS_SYSTEM
 
 #if PAL_HAS_THREAD
-    registerTest("Thread Test", threadTest);
-    registerTest("TLS Test", tlsTest);
-    registerTest("Mutex Test", mutexTest);
-    registerTest("Condvar Test", condvarTest);
+    // registerTest("Thread Test", threadTest);
+    // registerTest("TLS Test", tlsTest);
+    // registerTest("Mutex Test", mutexTest);
+    // registerTest("Condvar Test", condvarTest);
 #endif // PAL_HAS_THREAD
 
 #if PAL_HAS_VIDEO
@@ -33,6 +33,7 @@ int main(int argc, char** argv)
     registerTest("Icon Test", iconTest);
     registerTest("Cursor Test", cursorTest);
     registerTest("Input Window Test", inputWindowTest);
+    registerTest("System Cursor Test", systemCursorTest);
 #endif // PAL_HAS_VIDEO
 
 #if PAL_HAS_OPENGL